Replaced configuration file setup with config flow.
Replaced the existing configuration file setup with a more modern config flow setup. This also masks some of the complexity from users.
