+++ custom_components/dmp/alarm_control_panel.py
@@ -18,40 +18,62 @@ from homeassistant.components.alarm_control_panel.const import (
 )
 import homeassistant.helpers.config_validation as cv
 
-from .const import DOMAIN, LISTENER, CONF_AREA_NAME, CONF_AREA_ACCTNUM, CONF_AREA_NUMBER, CONF_AREA_DISARM_ZONE, CONF_AREA_HOME_ZONE, CONF_AREA_AWAY_ZONE
+from .const import (DOMAIN, LISTENER, CONF_AREA_NAME,
+                    CONF_PANEL_ACCOUNT_NUMBER, CONF_AREA_NUMBER,
+                    CONF_AREA_DISARM_ZONE, CONF_AREA_HOME_ZONE,
+                    CONF_AREA_AWAY_ZONE, CONF_AREAS)
 
 _LOGGER = logging.getLogger(__name__)
 
-PLATFORM_SCHEMA = vol.Schema(
-    {
-        vol.Required(CONF_AREA_NAME): cv.string,
-        vol.Required(CONF_AREA_ACCTNUM): cv.string,           
-        vol.Required(CONF_AREA_NUMBER): cv.string,
-        vol.Optional(CONF_AREA_DISARM_ZONE): cv.string,
-        vol.Optional(CONF_AREA_HOME_ZONE): cv.string,
-        vol.Optional(CONF_AREA_AWAY_ZONE): cv.string,
-    },
-    extra=vol.ALLOW_EXTRA,
-)
+# PLATFORM_SCHEMA = vol.Schema(
+#     {
+#         vol.Required(CONF_AREA_NAME): cv.string,
+#         vol.Required(CONF_AREA_ACCTNUM): cv.string,
+#         vol.Required(CONF_AREA_NUMBER): cv.string,
+#         vol.Optional(CONF_AREA_DISARM_ZONE): cv.string,
+#         vol.Optional(CONF_AREA_HOME_ZONE): cv.string,
+#         vol.Optional(CONF_AREA_AWAY_ZONE): cv.string,
+#     },
+#     extra=vol.ALLOW_EXTRA,
+# )
 
-async def async_setup_platform(hass, config, async_add_entities, discovery_info=None):
+
+async def async_setup_entry(hass, entry, async_add_entities,):
+    """Setup sensors from a config entry created in the integrations UI."""
+    hass.data.setdefault(DOMAIN, {})
+    config = hass.data[DOMAIN][entry.entry_id]
+    # Update our config to include new repos and remove those that have
+    # been removed.
+    # if config.options:
+    #     config.update(config.options)
     listener = hass.data[DOMAIN][LISTENER]
-    entity = [DMPArea(listener, config)]
-    async_add_entities(entity)
+    areas = [DMPArea(listener, area, config.get(CONF_PANEL_ACCOUNT_NUMBER))
+             for area in config[CONF_AREAS]]
+    async_add_entities(areas, update_before_add=True)
+
+
+# async def async_setup_platform(hass, config, async_add_entities,
+#                                discovery_info=None):
+#     listener = hass.data[DOMAIN][LISTENER]
+#     area = [DMPArea(listener, config)]
+#     async_add_entities(area)
 
 
 class DMPArea(AlarmControlPanelEntity):
-    def __init__(self, listener, config):
+    def __init__(self, listener, config, accountNum):
         self._listener = listener
         self._name = config.get(CONF_AREA_NAME)
-        self._account_number = config.get(CONF_AREA_ACCTNUM)
+        self._account_number = accountNum
         self._number = config.get(CONF_AREA_NUMBER)
         self._panel = listener.getPanels()[str(self._account_number)]
-        self._disarm_zone = config.get(CONF_AREA_DISARM_ZONE) or self._number[1:]
-        self._home_zone = config.get(CONF_AREA_HOME_ZONE) or self._number[1:]
-        self._away_zone = config.get(CONF_AREA_AWAY_ZONE) or self._number[1:]
-
-        areaObj = {"areaName": self._name, "areaNumber": str(self._number), "areaState": STATE_ALARM_DISARMED,}
+        self._disarm_zone = (config.get(CONF_AREA_DISARM_ZONE)
+                             or self._number[1:])
+        self._home_zone = (config.get(CONF_AREA_HOME_ZONE)
+                           or self._number[1:])
+        self._away_zone = (config.get(CONF_AREA_AWAY_ZONE)
+                           or self._number[1:])
+        areaObj = {"areaName": self._name, "areaNumber": str(self._number),
+                   "areaState": STATE_ALARM_DISARMED}
         self._panel.updateArea(str(self._number), areaObj)
 
     async def async_added_to_hass(self):
@@ -103,7 +125,7 @@ class DMPArea(AlarmControlPanelEntity):
     def unique_id(self):
         """Return unique ID"""
         return "dmp-%s-area-%s" % (self._account_number, self._number)
-    
+
     async def async_alarm_disarm(self, code=None):
         """Send disarm command."""
         await self._panel.connectAndSend('!O{},'.format(self._disarm_zone))
